[REFACTOR]: HomeTopArea 진행 단계 케이스별 파일 분리
742줄 단일 파일을 라우터와 4개 케이스 파일로 나누어
가독성을 개선하고 미사용 상수 제거
